requirements.txt
Django==3.0.5
asgiref==3.2.7
django-ckeditor==5.9.0
django-js-asset==1.2.2
pip==20.0.2
pytz==2019.3
setuptools==40.8.0
sqlparse==0.3.1
django-simple-captcha == 0.5.12
six == 1.14.0（上一项自动附带，无需手动安装）
Pillow == 7.1.2